==> Data Wrangling/Step 1 - Denoise.ipynb <==
import warnings
warnings.simplefilter('ignore')
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
from tqdm import tqdm
from dotenv import dotenv_values

import dask
import dask.dataframe as dd

config = dotenv_values("../.env")

def denoise(df):
    df['D_63'] = df['D_63'].apply(lambda t: {'CR':0, 'XZ':1, 'XM':2, 'CO':3, 'CL':4, 'XL':5}[t],
                                  meta=pd.DataFrame).astype(np.int8)
    df['D_64'] = df['D_64'].apply(lambda t: {np.nan:-1, 'O':0, '-1':1, 'R':2, 'U':3}[t],
                                  meta=pd.DataFrame).astype(np.int8)
    for col in tqdm(df.columns):
        if col not in ['customer_ID','S_2','D_63','D_64']:
            df[col] = np.floor(df[col]*100).astype(np.float32)
    return df

train = dd.read_csv(config["TRAIN_PATH"],blocksize=32e6)
train.head()

train = denoise(train)
train.head()

train.info()

train.to_parquet(config["WRANGLED_DATA"] + "denoised_train/", 
             name_function=lambda x : f"train-{x}.parquet")

del train

test = dd.read_csv(config["TEST_PATH"],blocksize=32e6)
test.head()

test = denoise(test)
test.head()

test.info()

test.to_parquet(config["WRANGLED_DATA"] + "denoised_test/", 
             name_function=lambda x : f"test-{x}.parquet")

## train_data.csv
__16gb -> 512mb__

## test_data.csv
__33gb -> 1gb__
