==> Data Wrangling/Step 3 - Standard Scaling & PCA.ipynb <==
import warnings
warnings.simplefilter('ignore')

import dask
import dask.dataframe as dd
import pandas as pd
import numpy as np
import os,gc

from tqdm import tqdm
from dotenv import dotenv_values
from dask_ml.decomposition import PCA
from dask_ml.preprocessing import StandardScaler

config = dotenv_values("../.env")

cat_ddf = dd.read_parquet(config["WRANGLED_DATA"] + "train_cat_feature.parquet", chunksize=512)
num_ddf = dd.read_parquet(config["WRANGLED_DATA"] + "train_num_feature.parquet", chunksize=512)
diff_ddf = dd.read_parquet(config["WRANGLED_DATA"] + "train_diff_feature.parquet", chunksize=512)
train_ddf = cat_ddf.merge(num_ddf, on="customer_ID").merge(diff_ddf, on="customer_ID").fillna(0)
train_ddf.shape[1]

cat_ddf = dd.read_parquet(config["WRANGLED_DATA"] + "test_cat_feature.parquet", chunksize=1024)
num_ddf = dd.read_parquet(config["WRANGLED_DATA"] + "test_num_feature.parquet", chunksize=1024)
diff_ddf = dd.read_parquet(config["WRANGLED_DATA"] + "test_diff_feature.parquet", chunksize=1024)
test_ddf = cat_ddf.merge(num_ddf, on="customer_ID").merge(diff_ddf, on="customer_ID").fillna(0)
test_ddf.shape[1]

res = [x for x in list(train_ddf.columns) + list(test_ddf.columns) 
       if x not in train_ddf.columns or x not in test_ddf.columns]
res

train_ddf = train_ddf.drop(res, axis=1)
len(train_ddf.columns)

all_cols = [x for x in list(train_ddf.columns) if x not in ["customer_ID", "target"]]
len(all_cols)

del cat_ddf, num_ddf, diff_ddf
gc.collect()

scaler = StandardScaler()
scaler

train_ddf = scaler.fit_transform(train_ddf[all_cols])
train_ddf.to_csv(config["WRANGLED_DATA"] + "scaled_train", name_function=lambda x: f"train-{x}.csv", index=False)

pca = PCA(n_components=128, svd_solver='randomized', whiten=True)
pca.fit(train_ddf.to_dask_array(lengths=True))

train_ddf = pca.transform(train_ddf.to_dask_array(lengths=True)).to_dask_dataframe()
train_ddf.to_csv(config["WRANGLED_DATA"] + "train_pca", name_function=lambda x: f"train-pca-{x}.csv", index=False)

test_ddf = scaler.transform(test_ddf[all_cols])
test_ddf.to_csv(config["WRANGLED_DATA"] + "scaled_test", name_function=lambda x: f"test-{x}.csv", index=False)

test_ddf = pca.transform(test_ddf.to_dask_array(lengths=True)).to_dask_dataframe()
test_ddf.to_csv(config["WRANGLED_DATA"] + "test_pca", name_function=lambda x: f"test-pca-{x}.csv", index=False)
